backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
pydantic==2.5.3
python-multipart==0.0.6
websockets==12.0
numpy>=1.24.0
pandas>=2.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
httpx>=0.25.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
